% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{reset_openai_config}
\alias{reset_openai_config}
\title{Clear session OpenAI configuration}
\usage{
reset_openai_config()
}
\value{
Invisibly \code{NULL}.
}
\description{
Removes any values set with \code{\link[=set_openai_config]{set_openai_config()}}, reverting to configuration
taken from a \code{.openai_config} file or environment variables.
}
\examples{
set_openai_config(model = "gpt-4o-mini")
reset_openai_config()

}
